feat(Adm): finishing
terminado os adapters

refs: PP-34

+++ App-Pizzaria/src/main/java/com/system/application/pizzaria/entity/Pagamento.java
@@ -4,6 +4,8 @@ import lombok.EqualsAndHashCode;
 import lombok.Getter;
 import lombok.Setter;
 
+import java.math.BigDecimal;
+
 @Getter
 @Setter
 @EqualsAndHashCode
@@ -11,6 +13,6 @@ public class Pagamento {
 
     private Integer idPagamento;
     private String funcionarioPagamento;
-    private Double salarioPagamento;
+    private BigDecimal salarioPagamento;
 
 }
