+++ App-Pizzaria/src/main/java/com/system/application/pizzaria/viewmodel/PagamentoVM.java
@@ -4,6 +4,8 @@ import lombok.EqualsAndHashCode;
 import lombok.Getter;
 import lombok.Setter;
 
+import java.math.BigDecimal;
+
 @Getter
 @Setter
 @EqualsAndHashCode
@@ -11,6 +13,6 @@ public class PagamentoVM {
 
     private Integer idPagamentoVM;
     private String funcionarioPagamentoVM;
-    private Double salarioPagamentoVM;
+    private BigDecimal salarioPagamentoVM;
 
 }
